Decouple news from ui tools required action in the welcome message
